Move Ripple main to Decorator approach
