+++ components/list/ListItem.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ClassNames from 'classnames';
 import FontIcon from '../font_icon';
 import ListItemContent from './ListItemContent';
-import RippleDecorator from '../ripple/RippleDecorator';
+import Ripple from '../ripple';
 import style from './style';
 
 class ListItem extends React.Component {
@@ -60,7 +60,7 @@ class ListItem extends React.Component {
   }
 }
 
-export default RippleDecorator({
+export default Ripple({
   className: style.ripple,
   centered: false
 })(ListItem);
