+++ components/switch/Thumb.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import RippleDecorator from '../ripple/RippleDecorator';
+import Ripple from '../ripple';
 import style from './style';
 
 const Check = ({children, onMouseDown}) => (
   <span role='thumb' className={style.thumb} onMouseDown={onMouseDown}>{children}</span>
 );
 
-export default RippleDecorator({
+export default Ripple({
   className: style.ripple,
   spread: 2.6,
   centered: true
